feat: Committer 기능 추가

+++ src/main/java/commity/web/repository/CommitterRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CommitterRepository extends JpaRepository<Committer, Long> {
+public interface CommitterRepository extends JpaRepository<Committer, Integer> {
 
     Optional<Committer> findByUsername(String username);
 }
